Shift persistence test and fixing a property type in ShiftPersistence

+++ LibrarySystem-Backend/src/main/java/ca/mcgill/ecse321/library/dao/ShiftRepository.java
@@ -8,9 +8,9 @@ import ca.mcgill.ecse321.library.models.HeadLibrarian;
 import ca.mcgill.ecse321.library.models.Librarian;
 import ca.mcgill.ecse321.library.models.Shift;
 
-public interface ShiftRepository extends CrudRepository<Shift, Integer>{
+public interface ShiftRepository extends CrudRepository<Shift, Long>{
     
-    Shift findShiftByShiftCode(int shiftCode); 
+    Shift findShiftByShiftCode(long shiftCode); 
     List<Shift> findShiftByLibrarians(Librarian librarians);
     List<Shift> findShiftByHeadLibrarian(HeadLibrarian headlibrarian);
